[Multi-label] Intersection matrix plot, tests, and doc
